formatting and small fixes
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/DWOProject/DWO-LMS/DWOServer/trunk@16519 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/main/java/fi/dwo/server/rest/SecuredUserScoContextManager.java
@@ -92,7 +92,6 @@ public class SecuredUserScoContextManager {
 			return Collections.emptyList();
 
 		List<PersistentScoContext> list = ScoContextManager.findEntities(parent);
-//sort?		
 		return list.stream().map((s)->s.buildDomScoContext()).sorted(new DomScoContextComparator()).collect(Collectors.toList());    	
     }
 
@@ -131,7 +130,7 @@ public class SecuredUserScoContextManager {
         PersistentSchool school = HasRoleUtilManager.getSchoolforHasRole(phr);
 		
 // profile match		
-		if ( pid != parent.getDwoProfileID())
+		if ( pid != parent.getDwoProfileID().longValue())
 			return null;
 // schools matches
 		if (parent.getSchoolID() != null) {
